Fix quiescence search as well

+++ src/bm/bm_search/move_gen.rs
@@ -270,8 +270,8 @@ impl QuiescenceSearchMoveGen {
                 piece_moves.to &= board.colors(!board.side_to_move());
                 for make_move in piece_moves {
                     let expected_gain =
-                        c_hist.get(board.side_to_move(), make_move.from, make_move.to)
-                            + search::see::<1>(&board, make_move) * 32;
+                        c_hist.get(board.side_to_move(), make_move.from, make_move.to) as i32
+                            + search::see::<1>(&board, make_move) as i32 * 32;
                     self.queue.push((make_move, expected_gain, None));
                 }
                 false
